Explain Project via README.md
I have added information about the project and explained in-depth
what frameworks the project is using to achieve its goal. Additionally,
I have added minor files (new project logo, yay!) to accompany
the README and installation of the project.

This will be the latest commit for a few months until I can find time
to continue investing in this project.